+++ examples/invoicing_app/lib/invoicing_app/sales/invoice_created_signal.ex
@@ -1,7 +1,7 @@
 defmodule InvoicingApp.Sales.InvoiceCreatedSignal do
   use Sea.Signal
 
-  emit_within(InvoicingApp.{Analytics, Customers, Inventory})
+  emit_to InvoicingApp.{Analytics, Customers, Inventory}
 
   defstruct [:customer_id, :product_id]